requirements.txt
crewai>=0.28.0
crewai[tools]>=0.28.0
python-dotenv>=1.0.0
PyGithub>=2.1.1
gitpython>=3.1.40
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
tiktoken>=0.5.0
requests>=2.31.0
flask>=3.0.0
